fix: libc++ buffer overflow in string_view ctor
Refs https://github.com/llvm/llvm-project/issues/61100.

Fixes a buffer overflow crash in std::string view constructor. This
happens as a result of a limitation of libc++'s implementation of the
string_view constructor. If string_view receives a pointer to a string start point,
and then a length as a size_t, the spec says that a size_t that exceeds the length
of the null terminated string found at pointer only be read up until the null terminator.

However, Chromium's implementation however does a hard check that this length is less than
or equal to static_cast<size_t>(std::numeric_limits<std::ptrdiff_t>::max()):
https://source.chromium.org/chromium/chromium/src/+/main:buildtools/third_party/libc++/trunk/include/string_view;drc=1718a75513d114e6b9aa4474e5c55d8dabee92fb;l=309

This doesn't break in Node.js right now because that hard assert is missing, but will
break in the next version of Xcode that's shipped and this Node.js too.